% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included-datasets.R
\docType{data}
\name{Canberra}
\alias{Canberra}
\title{Canberra weather data}
\format{
A tibble containing the downloaded and cleaned temperature data
\describe{
\item{Date}{The date of the temperature recording}
\item{Year}{The year of the temperature recording}
\item{Month}{The month of the temperature recording}
\item{Day}{The day of the temperature recording}
\item{Location}{The location (weather station name) of the temperature recording}
\item{Type}{The type of temperature recording (\code{Maximum} or \code{Minimum})}
}
}
\source{
\url{http://www.bom.gov.au/climate/data/}
}
\usage{
Canberra
}
\description{
A dataset containing daily temperatures for Canberra Airport (station number:
070351) downloaded from the Australian Bureau of Meteorology website:
\itemize{
\item \href{http://www.bom.gov.au/jsp/ncc/cdio/weatherData/av?p_nccObsCode=122&p_display_type=dailyDataFile&p_startYear=&p_c=&p_stn_num=070351}{Canberra Airport daily maximum temperatures}
\item \href{http://www.bom.gov.au/jsp/ncc/cdio/weatherData/av?p_nccObsCode=123&p_display_type=dailyDataFile&p_startYear=&p_c=&p_stn_num=070351}{Canberra Airport daily minimum temperatures}
}
}
\details{
Download date & time (approx.): 15 Mar 2024 @ 13PM ACDT
}
\keyword{datasets}
