% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_temperatures.R
\name{check_temperatures}
\alias{check_temperatures}
\title{Check if Australian temperature data are in the expected format}
\usage{
check_temperatures(data)
}
\arguments{
\item{data}{The object to be checked}
}
\description{
\code{check_temperatures()} checks if an object conforms to the format as per the dataset
returned from \code{\link{download_temperatures}} containing temperatures for
Australian weather stations.
}
\details{
An error is thrown if the object does not conform to the required format.
Otherwise the function returns nothing. Only limited error checking is
undertaken, such as checking for appropriate variable names and types.
Logical checks for aspects such as consistency within the dataset are not
performed.
}
\examples{
\dontrun{
# Download Adelaide temperature data (the default)
adelaide_data <- download_temperatures()

# Check the data for validity (nothing happens if valid)
check_temperatures(adelaide_data)
}

}
\seealso{
\code{\link{download_temperatures}}, \code{\link{plot_temperatures}}
}
