% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_temperatures.R
\name{download_temperatures}
\alias{download_temperatures}
\title{Download and clean Adelaide temperature data}
\usage{
download_temperatures()
}
\value{
A tibble of temperature data
}
\description{
\code{download_temperatures} downloads maximum and minimum daily temperatures
for Adelaide, cleans and reformats the data, and returns the combined data.
}
\details{
The data are from the \href{http://www.bom.gov.au/?ref=logo}{Australian
Bureau of Meteorology}. Temperature measurements are downloaded for all years
for the following stations in Adelaide, South Australia:
\itemize{
\item Adelaide (West Terrace / Ngayirdapira) (station number 23000) (there are no
data from 1980 to 2016); and
\item Adelaide Airport (station number 23034) (the data extend as far back as
1955).
}
}
