% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_temperatures.R
\name{get_locations}
\alias{get_locations}
\title{Get all locations from Australian temperature data}
\usage{
get_locations(data)
}
\arguments{
\item{data}{The tibble returned from \code{\link{download_temperatures}}
containing temperature data}
}
\value{
Sorted, unique locations from the temperature dataset
}
\description{
\code{get_locations()} returns all unique locations (weather stations) in the
temperature dataset returned by \code{\link{download_temperatures}}.
}
\examples{
\dontrun{
# Download Adelaide temperature data (the default)
adelaide_data <- download_temperatures()

# Extract locations
get_locations(adelaide_data)
}

}
\seealso{
\code{\link{download_temperatures}}, \code{\link{check_temperatures}}
}
