% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included-datasets.R
\docType{data}
\name{Hobart}
\alias{Hobart}
\title{Hobart weather data}
\format{
A tibble containing the downloaded and cleaned temperature data
\describe{
\item{Date}{The date of the temperature recording}
\item{Year}{The year of the temperature recording}
\item{Month}{The month of the temperature recording}
\item{Day}{The day of the temperature recording}
\item{Location}{The location (weather station name) of the temperature recording}
\item{Type}{The type of temperature recording (\code{Maximum} or \code{Minimum})}
}
}
\source{
\url{http://www.bom.gov.au/climate/data/}
}
\usage{
Hobart
}
\description{
A dataset containing daily temperatures for Hobart Airport West (station
number: 094008) downloaded from the Australian Bureau of Meteorology website:
\itemize{
\item \href{http://www.bom.gov.au/jsp/ncc/cdio/weatherData/av?p_nccObsCode=122&p_display_type=dailyDataFile&p_startYear=&p_c=&p_stn_num=094008}{Hobart Airport West daily maximum temperatures}
\item \href{http://www.bom.gov.au/jsp/ncc/cdio/weatherData/av?p_nccObsCode=123&p_display_type=dailyDataFile&p_startYear=&p_c=&p_stn_num=094008}{Hobart Airport West daily minimum temperatures}
}
}
\details{
Download date & time (approx.): 25 Mar 2024 @ 10PM ACDT
}
\keyword{datasets}
