% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_temperatures.R
\name{plot_temperatures}
\alias{plot_temperatures}
\title{Graph Australian temperature data}
\usage{
plot_temperatures(
  data,
  season,
  start_year = year(today()) - 30,
  end_year = year(today()),
  location = NULL,
  thresholds = NULL
)
}
\arguments{
\item{data}{The tibble returned from \code{\link{download_temperatures}}
containing temperature data}

\item{season}{One of \code{summer} or \code{winter}}

\item{start_year}{The starting year for the range of data to graph. Default
is 31 years ago from the current year}

\item{end_year}{The ending year for the range of data to graph. Default is
the current year}

\item{location}{The weather station in \code{data} to graph temperatures for.
If \code{NULL} provided (the default value), then the last location from
\code{get_locations(data)} is used}

\item{thresholds}{Three ascending numeric thresholds that define the
extreme temperatures to graph. If \code{NULL} provided (the default value), then
the default thresholds for summer (\code{c(30, 35, 40)}) and winter
(\code{c(0, 3, 5)}are provided}
}
\value{
A temperature graph
}
\description{
\code{plot_temperatures()} graphs maximum or minimum daily temperatures
for Australian weather stations.
}
\details{
Graphs for summer (maximum temperatures) or winter (minimum temperatures) can
be produced for a single weather station using the dataset returned from
\code{\link{download_temperatures}}.

If the starting and ending years provided are beyond the range of data, then
the earliest and latest available years will be used, respectively.

If at least one leap-year is present in the data, then February 29 will
appear in the graph for all years, otherwise it will not appear for any year.

Since the summer months used in the graphs (Nov thru Mar) span at least two
years, plotting only one summer season (such that \code{start_year == end_year})
will comprise at least two rows in the resulting graph.
}
\examples{
\dontrun{
# Download Adelaide temperature data (the default)
adelaide_data <- download_temperatures()

# Graph daily maximum temperatures
plot_temperatures(data = adelaide_data, season = 'summer')
}

}
\seealso{
\code{\link{download_temperatures}}, \code{\link{check_temperatures}},
\code{\link{get_locations}}
}
